created a base class for xyz renderers
(cherry picked from commit 3d9d40a1af043161346e858a389392efe3f0dd91)
