+++ chartfx-chart/src/main/java/io/fair_acc/chartfx/renderer/spi/ContourDataSetRenderer.java
@@ -77,7 +77,6 @@ import io.fair_acc.dataset.utils.ProcessingProfiler;
 public class ContourDataSetRenderer extends AbstractContourDataSetRendererParameter<ContourDataSetRenderer> implements Renderer {
     private static final Logger LOGGER = LoggerFactory.getLogger(ContourDataSetRenderer.class);
     private ContourDataSetCache localCache;
-    protected Axis zAxis;
     protected final ColorGradientBar gradientBar = new ColorGradientBar();
 
     private void drawContour(final GraphicsContext gc, final ContourDataSetCache lCache) {
@@ -288,58 +287,6 @@ public class ContourDataSetRenderer extends AbstractContourDataSetRendererParame
         return this;
     }
 
-    @Override
-    public void updateAxes() {
-        super.updateAxes();
-
-        // Check if there is a user-specified 3rd axis
-        if (zAxis == null) {
-            zAxis = tryGetZAxis(getAxes(), false);
-        }
-
-        // Fallback to one from the chart
-        if (zAxis == null) {
-            zAxis = tryGetZAxis(getChart().getAxes(), true);
-        }
-
-        // Fallback to adding one to the chart (to match behavior of X and Y)
-        if (zAxis == null) {
-            zAxis = createZAxis();
-            getChart().getAxes().add(zAxis);
-        }
-    }
-
-    private Axis tryGetZAxis(List<Axis> axes, boolean requireDimZ) {
-        Axis firstNonXY = null;
-        for (Axis axis : axes) {
-            if (axis != xAxis && axis != yAxis) {
-                // Prefer DIM_Z if possible
-                if (axis.getDimIndex() == DataSet.DIM_Z) {
-                    return axis;
-                }
-
-                // Potentially allow the first unused one
-                if (firstNonXY == null) {
-                    firstNonXY = axis;
-                }
-            }
-        }
-        return requireDimZ ? null : firstNonXY;
-    }
-
-    public static DefaultNumericAxis createZAxis() {
-        var zAxis = new DefaultNumericAxis("z-Axis");
-        zAxis.setAnimated(false);
-        zAxis.setSide(Side.RIGHT);
-        zAxis.setDimIndex(DataSet.DIM_Z);
-        return zAxis;
-    }
-
-    @Override
-    public boolean isUsingAxis(Axis axis) {
-        return super.isUsingAxis(axis) || axis == zAxis;
-    }
-
     /**
      * A rectangular color display that gets rendered next to the axis.
      * TODO: the layout currently requires the axis to be a child of a ChartPane.
@@ -496,14 +443,18 @@ public class ContourDataSetRenderer extends AbstractContourDataSetRendererParame
 
     }
 
+    @Override
     public void shiftZAxisToLeft() {
+        super.shiftZAxisToLeft();
         gradientBar.toBack();
         if (zAxis instanceof Node) {
             ((Node) zAxis).toBack();
         }
     }
 
+    @Override
     public void shiftZAxisToRight() {
+        super.shiftZAxisToRight();
         gradientBar.toFront();
         if (zAxis instanceof Node) {
             ((Node) zAxis).toFront();
